convert_service/requirements.txt
dnspython==2.7.0
ffmpeg-python==0.2.0
future==1.0.0
pika==1.3.2
pymongo==4.12.0
